+++ common/src/main/java/net/berserker_rpg/spell/BerserkerSpells.java
@@ -47,33 +47,6 @@ public class BerserkerSpells {
         return entry;
     }
 
-
-    private static Spell.Impact damageImpact(float coefficient, float knockback) {
-        var damage = new Spell.Impact();
-        damage.action = new Spell.Impact.Action();
-        damage.action.type = Spell.Impact.Action.Type.DAMAGE;
-        damage.action.damage = new Spell.Impact.Action.Damage();
-        damage.action.damage.spell_power_coefficient = coefficient;
-        damage.action.damage.knockback = knockback;
-        return damage;
-    }
-    private static void configureCooldown(Spell spell, float duration) {
-        if (spell.cost == null) {
-            spell.cost = new Spell.Cost();
-        }
-        spell.cost.cooldown = new Spell.Cost.Cooldown();
-        spell.cost.cooldown.duration = duration;
-    }
-    private static Spell.Impact createEffectImpact(Identifier effectId, float duration) {
-        var buff = new Spell.Impact();
-        buff.action = new Spell.Impact.Action();
-        buff.action.type = Spell.Impact.Action.Type.STATUS_EFFECT;
-        buff.action.status_effect = new Spell.Impact.Action.StatusEffect();
-        buff.action.status_effect.effect_id = effectId.toString();
-        buff.action.status_effect.duration = duration;
-        return buff;
-    }
-
     private static Spell modifierSpellBase() {
         var spell = new Spell();
         spell.range = 0;
@@ -91,28 +64,9 @@ public class BerserkerSpells {
 
         return spell;
     }
-    private static Spell.Impact.TargetModifier createImpactModifier(String entityType) {
-        var condition = new Spell.TargetCondition();
-        condition.entity_type = entityType;
-        var modifier = new Spell.Impact.TargetModifier();
-        modifier.conditions = List.of(condition);
-        return modifier;
-    }
-    private static void bleedingDeny(Spell.Impact impact) {
-        var modifier = createImpactModifier("#more_rpg_classes:bleeding_immune");
-        modifier.execute = TriState.DENY;
-        impact.target_modifiers = List.of(modifier);
-    }
-    private static Spell passiveSpellBase() {
-        var spell = new Spell();
-        spell.range = 0;
-        spell.tier = 8;
 
-        spell.type = Spell.Type.PASSIVE;
-        spell.passive = new Spell.Passive();
-        return spell;
-    }
-
+    public static final String BRUTE =  "brute";
+    public static final String FANATIC = "fanatic";
     ///MODIFIERS
     public static Entry improved_bloody_strike = add(improved_bloody_strike());
     private static Entry improved_bloody_strike() {
@@ -130,8 +84,7 @@ public class BerserkerSpells {
 
         return new Entry(id, spell, title, description, null, null);
     }
-    ///PASSIVES
-    ///ACTIVE SPELLS
+
     public static final Entry wild_rage = add(wild_rage());
     private static Entry wild_rage() {
         var id = Identifier.of(MOD_ID, "wild_rage");
@@ -142,6 +95,7 @@ public class BerserkerSpells {
         spell.range = 0;
         spell.tier = 2;
         spell.school = MoreSpellSchools.RAGE_MELEE;
+        spell.group = FANATIC;
         SpellTooltip.DescriptionMutator mutator = (args) -> {
             var modifier = effect.config().attributes().get(1);
             var modifier2 = effect.config().attributes().get(0);
@@ -182,14 +136,11 @@ public class BerserkerSpells {
         stashMeleeTrigger.target_override = Spell.Trigger.TargetSelector.CASTER;
         spell.deliver.stash_effect.triggers = List.of(stashMeleeTrigger);
 
-        var buff = createEffectImpact(effect.id, 10);
-        buff.action.status_effect.apply_mode = Spell.Impact.Action.StatusEffect.ApplyMode.ADD;
-        buff.action.status_effect.amplifier = 1;
-        buff.action.status_effect.amplifier_cap = 9;
+        var buff = SpellBuilder.Impacts.effectAdd(effect.id.toString(),10,1,9);
         buff.action.status_effect.refresh_duration = false;
         spell.impacts = List.of(buff);
 
-        configureCooldown(spell, 15);
+        SpellBuilder.Cost.cooldown(spell, 15);
         spell.cost.exhaust = 0.2F;
 
         return new Entry(id, spell, title, description, mutator,Book.BERSERKER);
@@ -206,14 +157,13 @@ public class BerserkerSpells {
         spell.range_mechanic = Spell.RangeMechanic.MELEE;
         spell.tier = 3;
         spell.school = MoreSpellSchools.RAGE_MELEE;
+        spell.group = FANATIC;
 
         spell.release.animation = PlayerAnimation.of("berserker_rpg:berserker_axe_both");
 
-        var buff = createEffectImpact(effect.id, 15);
+        var buff = SpellBuilder.Impacts.effectSet_ScaledAmplifier(
+                effect.id.toString(), 15,1,0.5F);
         buff.action.apply_to_caster = true;
-        buff.action.status_effect.apply_mode = Spell.Impact.Action.StatusEffect.ApplyMode.SET;
-        buff.action.status_effect.show_particles = false;
-        buff.action.status_effect.amplifier_power_multiplier = 0.5F;
         buff.particles = new ParticleBatch[]{
                 new ParticleBatch(
                         SpellEngineParticles.MagicParticles.get(
@@ -225,13 +175,8 @@ public class BerserkerSpells {
                         .color(Color.RAGE.toRGBA())
         };
 
-        var debuff = createEffectImpact(debuffEffect.id, 5);
-        bleedingDeny(debuff);
-        debuff.action.status_effect.apply_mode = Spell.Impact.Action.StatusEffect.ApplyMode.ADD;
-        debuff.action.status_effect.show_particles = false;
-        debuff.action.status_effect.amplifier = 1;
-        debuff.action.status_effect.amplifier_cap = 3;
-        debuff.action.status_effect.amplifier_cap_power_multiplier = 0.2F;
+        var debuff = SpellBuilder.Impacts.effectAdd_ScaledAmplifier(
+                debuffEffect.id.toString(), 5,1,0.2F);
         debuff.particles = new ParticleBatch[]{
                 new ParticleBatch(SpellEngineParticles.dripping_blood.id().toString(),
                         ParticleBatch.Shape.SPHERE, ParticleBatch.Origin.CENTER,
@@ -246,12 +191,12 @@ public class BerserkerSpells {
         spell.target.area.vertical_range_multiplier = 0.5F;
         spell.target.area.angle_degrees = 45;
 
-        var damage = damageImpact(0.65F, 1.0F);
+        var damage = SpellBuilder.Impacts.damage(0.65F, 1.0F);
         damage.sound = new Sound(BerserkerSounds.BLOODY_STRIKE.id());
 
         spell.impacts = List.of(debuff, damage, buff);
         
-        configureCooldown(spell, 13);
+        SpellBuilder.Cost.cooldown(spell, 13);
         spell.cost.exhaust = 0.3F;
         return new Entry(id, spell, title, description, null,Book.BERSERKER);
     }
@@ -266,6 +211,7 @@ public class BerserkerSpells {
         spell.range_mechanic = Spell.RangeMechanic.MELEE;
         spell.tier = 2;
         spell.school = MoreSpellSchools.RAGE_MELEE;
+        spell.group = BRUTE;
 
         spell.release.animation = PlayerAnimation.of("more_rpg_classes:two_handed_roar");
         spell.release.sound = new Sound(BerserkerSounds.OUTRAGE.id());
@@ -280,17 +226,17 @@ public class BerserkerSpells {
         spell.target.area.vertical_range_multiplier = 0.5F;
         spell.target.area.angle_degrees = 120;
 
-        var damage = damageImpact(0.5F, 0F);
+        var damage = SpellBuilder.Impacts.damage(0.5F, 0F);
         damage.sound = new Sound(BerserkerSounds.BLOODY_STRIKE.id());
 
         var pull = SpellBuilder.Impacts.pull(0.8F);
 
-        var debuff = createEffectImpact(effect.id, 6);
+        var debuff = SpellBuilder.Impacts.effectAdd_ScaledAmplifier(effect.id.toString(), 6,0,0.15F);
         debuff.action.status_effect.apply_mode = Spell.Impact.Action.StatusEffect.ApplyMode.SET;
 
         spell.impacts = List.of(damage, pull, debuff);
 
-        configureCooldown(spell, 16);
+        SpellBuilder.Cost.cooldown(spell, 16);
         spell.cost.exhaust = 0.3F;
         return new Entry(id, spell, title, description, null, Book.BERSERKER);
     }
@@ -303,6 +249,7 @@ public class BerserkerSpells {
         var spell = SpellBuilder.createSpellActive();
         spell.range = 0;
         spell.tier = 4;
+        spell.group = FANATIC;
         spell.school = MoreSpellSchools.RAGE_MELEE;
 
         spell.release.animation = PlayerAnimation.of("more_rpg_classes:two_handed_roar");
@@ -324,13 +271,13 @@ public class BerserkerSpells {
 
         spell.target.type = Spell.Target.Type.CASTER;
 
-        var buff = createEffectImpact(effect.id, 20);
+        var buff = SpellBuilder.Impacts.effectSet(effect.id.toString(), 12,0);
         buff.action.status_effect.apply_mode = Spell.Impact.Action.StatusEffect.ApplyMode.SET;
 
         spell.impacts = List.of(buff);
         spell.cost.exhaust = 0.3F;
 
-        configureCooldown(spell, 20);
+        SpellBuilder.Cost.cooldown(spell, 20);
         return new Entry(id, spell, title, description, null,Book.BERSERKER);
     }
     public static final Entry outrage = add(outrage());
@@ -344,6 +291,7 @@ public class BerserkerSpells {
         spell.range = 0;
         spell.range_mechanic = Spell.RangeMechanic.MELEE;
         spell.tier = 3;
+        spell.group = BRUTE;
         SpellTooltip.DescriptionMutator mutator = (args) -> {
             var modifier = effect.config().firstModifier();
             var bonus = SpellTooltip.bonus(modifier.value, modifier.operation);
@@ -376,13 +324,13 @@ public class BerserkerSpells {
 
         spell.target.type = Spell.Target.Type.CASTER;
 
-        var buff = createEffectImpact(effect.id, 10);
-        buff.action.status_effect.apply_mode = Spell.Impact.Action.StatusEffect.ApplyMode.SET;
-        buff.action.status_effect.amplifier = 0;
+        var buff = SpellBuilder.Impacts.effectSet(effect.id.toString(), 10,0);
 
-        spell.impacts = List.of(buff);
+        var cleanse = SpellBuilder.Impacts.effectCleanse();
 
-        configureCooldown(spell, 35);
+        spell.impacts = List.of(buff, cleanse);
+
+        SpellBuilder.Cost.cooldown(spell, 35);
         spell.cost.exhaust = 0.3F;
         return new Entry(id, spell, title, description, mutator,Book.BERSERKER);
     }
@@ -397,13 +345,14 @@ public class BerserkerSpells {
         spell.range_mechanic = Spell.RangeMechanic.MELEE;
         spell.range = 0.5F;
         spell.tier = 4;
+        spell.group = BRUTE;
 
         spell.target.type = Spell.Target.Type.AIM;
         spell.target.aim = new Spell.Target.Aim();
         spell.target.aim.sticky = false;
         spell.target.aim.required = true;
 
-        spell.release.animation = PlayerAnimation.of("berserker_rpg:decapitate_charge");
+        spell.release.animation = PlayerAnimation.of("berserker_rpg:decapitate_release");
         spell.release.sound = new Sound(BerserkerSounds.DECAPITATE_RELEASE.id());
 
         var custom = new Spell.Impact();
@@ -416,199 +365,9 @@ public class BerserkerSpells {
 
         spell.impacts = List.of(custom);
 
-        configureCooldown(spell, 30);
+        SpellBuilder.Cost.cooldown(spell, 30);
         spell.cost.exhaust = 1.0F;
         spell.cost.durability = 1;
         return new Entry(id, spell, title, description, null, Book.BERSERKER);
-    }
-    public static final Entry rumbling_swing = add(rumbling_swing());
-    private static Entry rumbling_swing() {
-        var id = Identifier.of(MOD_ID, "rumbling_swing");
-        var title = "Rumbling Swing";
-        var description = "Jumps lightning fast, dealing {damage_0} physical- and {damage_1} lightning-damage";
-        var spell = SpellBuilder.createSpellActive();
-        spell.school = SpellSchools.LIGHTNING;
-        spell.range = 10;
-        spell.tier = 5;
-
-        spell.target.type = Spell.Target.Type.AIM;
-        spell.target.aim = new Spell.Target.Aim();
-        spell.target.aim.sticky = false;
-        spell.target.aim.required = true;
-
-        spell.release.animation = PlayerAnimation.of("berserker_rpg:rumbling_swing");
-        spell.release.sound = new Sound("entity.player.attack.sweep");
-        spell.release.particles = new ParticleBatch[]{
-                new ParticleBatch("electric_spark",
-                        ParticleBatch.Shape.CIRCLE, ParticleBatch.Origin.CENTER,
-                        15, 0.5F, 3.0F)
-        };
-
-        var teleport = new Spell.Impact();
-        teleport.action = new Spell.Impact.Action();
-        teleport.action.type = Spell.Impact.Action.Type.TELEPORT;
-        teleport.action.teleport = new Spell.Impact.Action.Teleport();
-        teleport.action.teleport.mode = Spell.Impact.Action.Teleport.Mode.BEHIND_TARGET;
-        teleport.action.teleport.intent = SpellTarget.Intent.HARMFUL;
-        teleport.action.teleport.behind_target = new Spell.Impact.Action.Teleport.BehindTarget();
-        teleport.action.teleport.behind_target.distance = 1.5F;
-        teleport.action.teleport.depart_particles = new ParticleBatch[]{
-                new ParticleBatch("cloud",
-                        ParticleBatch.Shape.SPHERE, ParticleBatch.Origin.FEET,
-                        20, 0.05F, 0.1F)
-                        .invert()
-                        .preSpawnTravel(15)
-        };
-        teleport.action.teleport.arrive_particles = new ParticleBatch[]{
-                new ParticleBatch("poof",
-                        ParticleBatch.Shape.SPHERE, ParticleBatch.Origin.FEET,
-                        10, 0.05F, 0.1F)
-                        .preSpawnTravel(2)
-        };
-
-        var damage = damageImpact(0.8F, 1.0F);
-        damage.attribute = "minecraft:generic.attack_damage";
-        damage.particles = new ParticleBatch[]{
-                new ParticleBatch("berserker_rpg:small_thunder",
-                        ParticleBatch.Shape.PIPE, ParticleBatch.Origin.CENTER,
-                        1, 0.1F, 0.15F)
-        };
-        damage.sound = new Sound("entity.lightning_bolt.impact");
-
-        var lightningDamage = new Spell.Impact();
-        lightningDamage.school = SpellSchools.LIGHTNING;
-        lightningDamage.action = new Spell.Impact.Action();
-        lightningDamage.action.type = Spell.Impact.Action.Type.DAMAGE;
-        lightningDamage.action.damage = new Spell.Impact.Action.Damage();
-        lightningDamage.action.damage.spell_power_coefficient = 0.5F;
-
-        spell.impacts = List.of(teleport, damage, lightningDamage);
-
-        configureCooldown(spell, 25);
-        spell.cost.exhaust = 1.0F;
-        spell.cost.durability = 1;
-        return new Entry(id, spell, title, description, null,null);
-    }
-    public static final Entry nordic_storm = add(nordic_storm());
-    private static Entry nordic_storm() {
-        var id = Identifier.of(MOD_ID, "nordic_storm");
-        var title = "Nordic Storm";
-        var description = "Spinning attack, that deals {damage_0} physical- and {damage_1} frost-damage on targets per second.";
-        var spell = SpellBuilder.createSpellActive();
-        spell.school = SpellSchools.FROST;
-        spell.range = 0;
-        spell.range_mechanic = Spell.RangeMechanic.MELEE;
-        spell.tier = 5;
-
-        spell.active.cast = new Spell.Active.Cast();
-        spell.active.cast.duration = 0.5F;
-        spell.active.cast.movement_speed = 1.2F;
-        spell.active.cast.animation = PlayerAnimation.of("berserker_rpg:nordic_storm");
-        spell.active.cast.sound = new Sound(Identifier.of("spell_engine:generic_frost_impact"), 0);
-        spell.active.cast.start_sound = new Sound("entity.player.breath");
-        spell.active.cast.type = Spell.Active.Cast.Type.CHANNEL;
-        spell.active.cast.channel = new Spell.Active.Cast.Channel();
-        spell.active.cast.channel.ticks = 5;
-        spell.active.cast.particles = new ParticleBatch[]{
-                new ParticleBatch("sweep_attack",
-                        ParticleBatch.Shape.PIPE, ParticleBatch.Origin.CENTER,
-                        1, 0.3F, 3.0F),
-                new ParticleBatch("spell_engine:frost_shard",
-                        ParticleBatch.Shape.SPHERE, ParticleBatch.Origin.CENTER,
-                        1, 0.3F, 3.0F)
-        };
-
-        spell.target.type = Spell.Target.Type.AREA;
-        spell.target.area = new Spell.Target.Area();
-        spell.target.area.distance_dropoff = Spell.Target.Area.DropoffCurve.SQUARED;
-        spell.target.area.angle_degrees = 360;
-        spell.target.area.vertical_range_multiplier = 0.5F;
-
-        var damage = damageImpact(0.8F, 0F);
-        damage.attribute = "minecraft:generic.attack_damage";
-        var freezeHurts = new Spell.Impact.TargetModifier();
-        var freezeHurtsCond = new Spell.TargetCondition();
-        freezeHurtsCond.entity_type = "#minecraft:freeze_hurts_extra_types";
-        freezeHurts.conditions = List.of(freezeHurtsCond);
-        freezeHurts.modifier = new Spell.Impact.Modifier();
-        freezeHurts.modifier.power_multiplier = 0.3F;
-        var freezeImmune = new Spell.Impact.TargetModifier();
-        var freezeImmuneCond = new Spell.TargetCondition();
-        freezeImmuneCond.entity_type = "#minecraft:freeze_immune_entity_types";
-        freezeImmune.conditions = List.of(freezeImmuneCond);
-        freezeImmune.modifier = new Spell.Impact.Modifier();
-        freezeImmune.modifier.power_multiplier = -0.3F;
-        damage.target_modifiers = List.of(freezeHurts, freezeImmune);
-        damage.particles = new ParticleBatch[]{
-                new ParticleBatch("spell_engine:frost_hit",
-                        ParticleBatch.Shape.SPHERE, ParticleBatch.Origin.LAUNCH_POINT,
-                        1, 0.5F, 3.0F)
-        };
-        damage.sound = new Sound(Identifier.of("item.axe.strip"), 2);
-
-        var frosted = createEffectImpact(Identifier.of("more_rpg_classes", "frosted"), 8);
-        frosted.action.status_effect.apply_mode = Spell.Impact.Action.StatusEffect.ApplyMode.ADD;
-        frosted.action.status_effect.amplifier = 1;
-        frosted.action.status_effect.amplifier_cap = 4;
-        frosted.action.status_effect.show_particles = false;
-        var frostedDeny = new Spell.Impact.TargetModifier();
-        var frostedDenyCond = new Spell.TargetCondition();
-        frostedDenyCond.entity_type = "#minecraft:freeze_immune_entity_types";
-        frostedDeny.conditions = List.of(frostedDenyCond);
-        frostedDeny.execute = TriState.DENY;
-        frosted.target_modifiers = List.of(frostedDeny);
-
-        spell.impacts = List.of(damage, frosted);
-
-        configureCooldown(spell, 10);
-        spell.cost.cooldown.proportional = true;
-        spell.cost.exhaust = 1.0F;
-        spell.cost.durability = 5;
-        return new Entry(id, spell, title, description, null, null);
-    }
-    /// WEAPON SKILLS
-    public static Entry decapitate = add(decapitate());
-    private static Entry decapitate() {
-        var id = Identifier.of(MOD_ID, "decapitate");
-        var title = "Decapitate";
-        var description = "Delivers a heavy blow with forward momentum that disables shield and item usage of target.";
-        var spell = SpellBuilder.createSpellActive();
-        spell.tier = 1;
-        spell.school = MoreSpellSchools.RAGE_MELEE;
-        spell.range = 0.0F;
-        spell.range_mechanic = Spell.RangeMechanic.MELEE;
-
-        SpellBuilder.Casting.cast(spell, 0.75F);
-        spell.active.cast.animation = PlayerAnimation.of("berserker_rpg:decapitate_charge");
-        spell.active.cast.animation.speed = 1.5F;
-        spell.active.cast.animation_pitch = false;
-        spell.release.sound = new Sound(BerserkerSounds.DECAPITATE_RELEASE.id());
-
-        SpellBuilder.Target.none(spell);
-
-        var cut_1 = new Spell.Delivery.Melee.Attack();
-        cut_1.attack_speed_multiplier = 1.25F;
-        cut_1.delay = 0.1F;
-        cut_1.hitbox = new Spell.Delivery.Melee.HitBox();
-        cut_1.hitbox.arc = 180;
-        cut_1.hitbox.height = 0.7F;
-        cut_1.hitbox.roll = 15F;
-        cut_1.damage_bonus = 0.25F;
-        cut_1.forward_momentum = 1.5F;
-        cut_1.swing_sound = new Sound(BerserkerSounds.DECAPITATE_SWING.id());
-        cut_1.impact_sound = new Sound(BerserkerSounds.DECAPITATE_IMPACT.id());
-        cut_1.impact_sound_cap = 1;
-        cut_1.animation = PlayerAnimation.of("berserker_rpg:decapitate_release");
-
-        SpellBuilder.Deliver.melee(spell, List.of(cut_1));
-
-        var disrupt = SpellBuilder.Impacts.disrupt(true, 2F);
-        spell.impacts = List.of(disrupt);
-
-        SpellBuilder.Cost.cooldownGroupWeapon(spell);
-        SpellBuilder.Cost.cooldown(spell, 15);
-        spell.cost.cooldown.attempt_duration = 1F;
-
-        return new Entry(id, spell, title, description, null, null);
     }
 }